+++ src/FluxFlow.Components.Http/HttpResultKinds.cs
@@ -1,0 +1,8 @@
+namespace FluxFlow.Components.Http;
+
+public static class HttpResultKinds
+{
+    public const string Response = "Response";
+
+    public const string Error = "Error";
+}